+++ client/src/components/Home/Home.jsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import '../Home/Home.css';
-
+import SwipeableTextMobileStepper from '../Carousel';
 import animal1 from '../Images/cat.jpg';
 import animal2 from '../Images/corgi.jpg';
 import animal3 from '../Images/pitbull.jpg';
@@ -10,149 +10,10 @@ const Home = () => {
         <div className="home">
             <h1>Make A Difference!</h1>
             <div className="animal-images">
-                <a href="/animal1" className="animal-link">
-                    <img className="animal-image" src={animal1} alt="Animal 1" />
-                    <span className="tooltip">Adopt</span>
-                </a>
-                <a href="/animal2" className="animal-link">
-                    <img className="animal-image" src={animal2} alt="Animal 2" />
-                    <span className="tooltip">Foster</span>
-                </a>
-                <a href="/animal3" className="animal-link">
-                    <img className="animal-image" src={animal3} alt="Animal 3" />
-                    <span className="tooltip">Volunteer</span>
-                </a>
+              <SwipeableTextMobileStepper/>
             </div>
         </div>
     );
 };
-// start of carousel
-import * as React from 'react';
-import { useTheme } from '@mui/material/styles';
-import Box from '@mui/material/Box';
-import MobileStepper from '@mui/material/MobileStepper';
-import Paper from '@mui/material/Paper';
-import Typography from '@mui/material/Typography';
-import Button from '@mui/material/Button';
-import KeyboardArrowLeft from '@mui/icons-material/KeyboardArrowLeft';
-import KeyboardArrowRight from '@mui/icons-material/KeyboardArrowRight';
-import SwipeableViews from 'react-swipeable-views';
-import { autoPlay } from 'react-swipeable-views-utils';
-
-const AutoPlaySwipeableViews = autoPlay(SwipeableViews);
-
-const images = [
-    {
-        label: '',
-        imgPath:
-            ' ',
-    },
-    {
-        label: '',
-        imgPath:
-            '',
-    },
-    {
-        label: '',
-        imgPath:
-            '   ',
-    },
-    {
-        label: ' ',
-        imgPath:
-            '',
-    },
-];
-
-function SwipeableTextMobileStepper() {
-    const theme = useTheme();
-    const [activeStep, setActiveStep] = React.useState(0);
-    const maxSteps = images.length;
-
-    const handleNext = () => {
-        setActiveStep((prevActiveStep) => prevActiveStep + 1);
-    };
-
-    const handleBack = () => {
-        setActiveStep((prevActiveStep) => prevActiveStep - 1);
-    };
-
-    // Removed TypeScript type annotation for compatibility with JSX
-    const handleStepChange = (step) => {
-        setActiveStep(step);
-    };
-
-    return (
-        <Box sx={{ maxWidth: 400, flexGrow: 1 }}>
-            <Paper
-                square
-                elevation={0}
-                sx={{
-                    display: 'flex',
-                    alignItems: 'center',
-                    height: 50,
-                    pl: 2,
-                    bgcolor: 'background.default',
-                }}
-            >
-                <Typography>{images[activeStep].label}</Typography>
-            </Paper>
-            <AutoPlaySwipeableViews
-                axis={theme.direction === 'rtl' ? 'x-reverse' : 'x'}
-                index={activeStep}
-                onChangeIndex={handleStepChange}
-                enableMouseEvents
-            >
-                {images.map((step, index) => (
-                    <div key={step.label}>
-                        {Math.abs(activeStep - index) <= 2 ? (
-                            <Box
-                                component="img"
-                                sx={{
-                                    height: 255,
-                                    display: 'block',
-                                    maxWidth: 400,
-                                    overflow: 'hidden',
-                                    width: '100%',
-                                }}
-                                src={step.imgPath}
-                                alt={step.label}
-                            />
-                        ) : null}
-                    </div>
-                ))}
-            </AutoPlaySwipeableViews>
-            <MobileStepper
-                steps={maxSteps}
-                position="static"
-                activeStep={activeStep}
-                nextButton={
-                    <Button
-                        size="small"
-                        onClick={handleNext}
-                        disabled={activeStep === maxSteps - 1}
-                    >
-                        Next
-                        {theme.direction === 'rtl' ? (
-                            <KeyboardArrowLeft />
-                        ) : (
-                            <KeyboardArrowRight />
-                        )}
-                    </Button>
-                }
-                backButton={
-                    <Button size="small" onClick={handleBack} disabled={activeStep === 0}>
-                        {theme.direction === 'rtl' ? (
-                            <KeyboardArrowRight />
-                        ) : (
-                            <KeyboardArrowLeft />
-                        )}
-                        Back
-                    </Button>
-                }
-            />
-        </Box>
-    );
-}
 
 export default Home;
