+++ package.json
@@ -18,7 +18,8 @@
     "gsap": "^3.12.5",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
-    "react-icons": "^5.2.1"
+    "react-icons": "^5.2.1",
+    "react-swipeable-views-v18": "^1.1.27"
   },
   "devDependencies": {
     "@types/react": "^18.0.27",
